requirements.txt
# Ansible
ansible==2.9.*

# Git versioning
pbr!=2.1.0,>=2.0.0 # Apache-2.0

# ini parsing
oslo.config>=5.2.0 # Apache-2.0

# YAML parsing
PyYAML>=3.12 # MIT

# templating
Jinja2>=2.10 # BSD License (3 clause)

# Ansible's json_query
jmespath>=0.9.3 # MIT
